Fix CloudConfigParams in Servers
CloudConfigParams in `CreateServerParams` was not properly tested when
first introduced. The assumption that a string would do, caused issues
with the json.Encode further down.

Using `map[string]any` solves this by allowing the user to declare a
custom map to be sent to the API. **This requires an update to Go
1.18.**

This commit fixes the `CloudConfigParams` in `CreateServerParams` and
properly implement the missing `CloudConfigParams` in
`PreviewCloudConfigParams` function.

* Added examples
* Added tests
* Update go.mod for 1.18 syntax
* Bump imdario/mergo to v0.3.13